homelab_airflow/dags/spotify_tracks/requirements.txt
apache-airflow
psycopg2
sqlalchemy==1.4.52
pymongo
redis
sqlmodel

spotipy
tqdm
docker
